+++ coordinate.h
@@ -1,13 +1,13 @@
 class Coordinate{
   private:
-    float x;
-    float y;
+    double x;
+    double y;
 
   public:
-    Coordinate(float x, float y);
-    float get_x(){ return x; }
-    float get_y(){ return y; }
-    void set_x(float new_x){ x = new_x; }
-    void set_y(float new_y){ y = new_y; }
-    static float distance(Coordinate* coordinate1, Coordinate* coordinate2);
+    Coordinate(double x, double y);
+    double get_x(){ return x; }
+    double get_y(){ return y; }
+    void set_x(double new_x){ x = new_x; }
+    void set_y(double new_y){ y = new_y; }
+    static double distance(Coordinate* coordinate1, Coordinate* coordinate2);
 };
