+++ road.h
@@ -1,11 +1,11 @@
 class Road{
   private:
-    float width;
+    double width;
     Coordinate* start_position;
     Coordinate* end_position;
     void draw_road();
     void Create();
   public:
-    Road(float width, Coordinate* start_position, Coordinate* end_position);
+    Road(double width, Coordinate* start_position, Coordinate* end_position);
     ~Road();
 };
