+++ velocity.cpp
@@ -1,5 +1,5 @@
 #include "velocity.h"
 
-Velocity::Velocity(float init_speed, char init_direction) : speed(init_speed), direction(init_direction)
+Velocity::Velocity(double init_speed, char init_direction) : speed(init_speed), direction(init_direction)
 {
 }
