+++ velocity.h
@@ -1,11 +1,11 @@
 class Velocity{
   private:
-    float speed;
+    double speed;
     char direction;
 
   public:
-    Velocity(float init_speed, char init_direction);
-    float get_speed(){ return speed; }
+    Velocity(double init_speed, char init_direction);
+    double get_speed(){ return speed; }
     char get_direction(){ return direction; }
-    void set_speed(float new_speed){ speed = new_speed; }
+    void set_speed(double new_speed){ speed = new_speed; }
 };
